requirements.txt
numpy==2.3.3
PyOpenGL==3.1.10
pyglet==2.1.13
Pillow==11.2.1

